requirements.txt
vanna[google,fastapi]>=2.0.0
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
pydantic
